requirements.txt
flask
tensorflow
keras
numpy
pandas
matplotlib
opencv-python
Pillow
scikit-learn